moved the lbd computer outside of the nogood folder, and deleted the nogood folder
