+++ pumpkin-solver/src/engine/constraint_satisfaction_solver.rs
@@ -15,7 +15,6 @@ use super::conflict_analysis::ConflictAnalysisContext;
 use super::conflict_analysis::LearnedNogood;
 use super::conflict_analysis::NoLearningResolver;
 use super::conflict_analysis::SemanticMinimiser;
-use super::nogoods::Lbd;
 use super::propagation::contexts::PropagationContextWithTrailedValues;
 use super::propagation::store::PropagatorStore;
 use super::propagation::PropagatorId;
@@ -23,6 +22,7 @@ use super::solver_statistics::SolverStatistics;
 use super::termination::TerminationCondition;
 use super::variables::IntegerVariable;
 use super::variables::Literal;
+use super::Lbd;
 use super::ResolutionResolver;
 use super::TrailedValues;
 use crate::basic_types::moving_averages::MovingAverage;
