+++ pumpkin-solver/src/engine/mod.rs
@@ -2,7 +2,7 @@ pub(crate) mod conflict_analysis;
 pub(crate) mod constraint_satisfaction_solver;
 pub(crate) mod cp;
 mod debug_helper;
-pub(crate) mod nogoods;
+mod literal_block_distance;
 pub(crate) mod predicates;
 mod restart_strategy;
 mod solver_statistics;
@@ -17,6 +17,7 @@ pub(crate) use cp::*;
 pub(crate) use debug_helper::DebugDyn;
 pub(crate) use debug_helper::DebugHelper;
 pub(crate) use domain_events::DomainEvents;
+pub(crate) use literal_block_distance::Lbd;
 pub use restart_strategy::RestartOptions;
 pub(crate) use restart_strategy::RestartStrategy;
 pub(crate) use solver_statistics::SolverStatistics;
