--- pumpkin-solver/src/engine/nogoods/mod.rs
@@ -1,3 +1,0 @@
-mod literal_block_distance;
-
-pub(crate) use literal_block_distance::Lbd;
